Refactor nzbToMediaUtil -> utils
