+++ core/__init__.py
@@ -45,7 +45,7 @@ from six.moves import reload_module
 from core import logger, main_db, version_check, databases, transcoder
 from core.auto_process import Comic, Game, Movie, Music, TV
 from core.configuration import config
-from core.nzbToMediaUtil import (
+from core.utils import (
     RunningProcess, wake_up, category_search, clean_dir, clean_dir, copy_link,
     create_torrent_class, extract_files, flatten, get_dirs, get_download_info,
     list_media_files, make_dir, parse_args, pause_torrent, remove_torrent,
