+++ core/auto_process/games.py
@@ -7,7 +7,7 @@ import requests
 
 import core
 from core import logger
-from core.nzbToMediaUtil import convert_to_ascii, server_responding
+from core.utils import convert_to_ascii, server_responding
 
 requests.packages.urllib3.disable_warnings()
