+++ core/auto_process/movies.py
@@ -9,7 +9,7 @@ import requests
 import core
 from core import logger, transcoder
 from core.scene_exceptions import process_all_exceptions
-from core.nzbToMediaUtil import convert_to_ascii, find_download, find_imdbid, import_subs, list_media_files, remote_dir, report_nzb, remove_dir, server_responding
+from core.utils import convert_to_ascii, find_download, find_imdbid, import_subs, list_media_files, remote_dir, remove_dir, report_nzb, server_responding
 
 requests.packages.urllib3.disable_warnings()
