+++ core/auto_process/tv.py
@@ -12,7 +12,7 @@ import core
 from core import logger, transcoder
 from core.forks import auto_fork
 from core.scene_exceptions import process_all_exceptions
-from core.nzbToMediaUtil import convert_to_ascii, flatten, import_subs, list_media_files, remote_dir, report_nzb, remove_dir, server_responding
+from core.utils import convert_to_ascii, flatten, import_subs, list_media_files, remote_dir, remove_dir, report_nzb, server_responding
 
 requests.packages.urllib3.disable_warnings()
